refactor(frontend): split Compiler into compiler/ translation units
Move the two cohesive blocks the plan calls out out of ir/Compiler.cpp:

  compiler/ClassMeta.cpp (~213 lines)
    registerClassMeta / compileClassBodies / methodSignature

  compiler/ForIn.cpp (~210 lines)
    compileForInCounter / compileForInIterable
    emitIterableLength / emitIterableElement

Compiler.cpp drops from 1603 to ~1206 lines and keeps every visit() method
plus the emit*/add* helpers that back them.

As with SemanticAnalyzer and Parser, only method bodies are relocated: the
class layout, the call graph and each `this->` access are untouched, so the
split is behaviour-preserving by construction.

suspendLoops() becomes an inline member in Compiler.h. Its return type is a
ScopeGuard over an unnamed lambda, which no caller in another translation
unit can spell or deduce; defining it in the header keeps the body identical
to the previous out-of-line definition while making it deducible at every
call site (Compiler.cpp and compiler/ClassMeta.cpp both use it).

The two qualifiedName overloads stay in an anonymous namespace: unlike the
shared helpers in semantic/Helpers.h and parser/ParserTokens.h, they are
only ever called from Compiler.cpp, so they keep internal linkage and add
no symbols to the unity build.

Verification: 429 tests pass (same 5 pre-existing failures as the baseline
branch), and the 52-script bytecode dump is byte-identical to the baseline.
No new warnings under -Wall -Wextra -Wshadow.
