+++ src/LOICollectionA/frontend/ir/Compiler.cpp
@@ -13,6 +13,7 @@
 #include "LOICollectionA/frontend/ir/Compiler.h"
 
 namespace LOICollection::frontend::ir {
+
     namespace {
         std::string qualifiedName(const VariableNode& node) {
             return node.isStaticField ? node.staticClassName + "::" + node.name : node.name;
@@ -26,14 +27,6 @@ namespace LOICollection::frontend::ir {
     }
 
     Compiler::Compiler(DiagnosticEngine& diag) : current(std::ref(chunk)), diagnostics(diag) {}
-
-    auto Compiler::suspendLoops() {
-        auto saved = std::exchange(this->loopStack, {});
-
-        return make_scope_guard([this, saved = std::move(saved)]() mutable {
-            this->loopStack = std::move(saved);
-        });
-    }
 
     BytecodeChunk Compiler::compile(ASTNode& root) {
         this->pushScope(false, 0);
@@ -287,200 +280,6 @@ namespace LOICollection::frontend::ir {
             this->compileForInCounter(node, uid);
         else
             this->compileForInIterable(node, uid, *protocol);
-    }
-
-    void Compiler::emitIterableLength(
-        const IterableProtocol& protocol, int seqSlot, const SourceLocation& loc
-    ) {
-        this->current.get().emit(OpCode::LOAD_SLOT, seqSlot, loc);
-
-        if (protocol.shape == IterableShape::Convention) {
-            if (ClassCall::getInstance().isRegistered(protocol.className)) {
-                int metaIdx = this->addNativeCall(protocol.className, std::string(lengthMethod), 0);
-                this->current.get().emit(OpCode::CALL_NATIVE_METHOD, metaIdx, loc);
-
-                return;
-            }
-
-            int metaIdx = this->addByNameCall(std::string(lengthMethod), 0);
-            this->current.get().emit(OpCode::CALL_METHOD_BY_NAME, metaIdx, loc);
-
-            return;
-        }
-
-        this->current.get().emit(OpCode::LOAD_LEN, 0, loc);
-    }
-
-    void Compiler::emitIterableElement(
-        const IterableProtocol& protocol, int seqSlot, int idxSlot, const SourceLocation& loc
-    ) {
-        if (protocol.shape == IterableShape::Convention) {
-            if (ClassCall::getInstance().isRegistered(protocol.className)) {
-                this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, loc);
-                this->current.get().emit(OpCode::LOAD_SLOT, seqSlot, loc);
-
-                int metaIdx = this->addNativeCall(protocol.className, std::string(elementMethod), 1);
-                this->current.get().emit(OpCode::CALL_NATIVE_METHOD, metaIdx, loc);
-
-                return;
-            }
-
-            this->current.get().emit(OpCode::LOAD_SLOT, seqSlot, loc);
-            this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, loc);
-
-            int metaIdx = this->addByNameCall(std::string(elementMethod), 1);
-            this->current.get().emit(OpCode::CALL_METHOD_BY_NAME, metaIdx, loc);
-
-            return;
-        }
-
-        this->current.get().emit(OpCode::LOAD_SLOT, seqSlot, loc);
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, loc);
-        this->current.get().emit(OpCode::LOAD_INDEX, 0, loc);
-    }
-
-    void Compiler::compileForInIterable(ForInNode& node, size_t uid, const IterableProtocol& protocol) {
-        const int seqSlot = this->declareSlot("__forin_seq_" + std::to_string(uid));
-        const int idxSlot = this->declareSlot("__forin_idx_" + std::to_string(uid));
-        const int elemSlot = this->declareSlot(node.elementVar);
-        const int indexSlot = node.hasIndexVar ? this->declareSlot(node.indexVar) : -1;
-
-        this->compileValue(*node.iterable, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, seqSlot, node.loc);
-
-        int zeroIdx = this->addConstant(0);
-        this->current.get().emit(OpCode::PUSH_INT, zeroIdx, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, idxSlot, node.loc);
-
-        size_t loopStart = this->current.get().currentIP();
-
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-        this->emitIterableLength(protocol, seqSlot, node.loc);
-        this->current.get().emit(OpCode::CMP_LT, 0, node.loc);
-
-        size_t jmpFalseIdx = this->current.get().emit(OpCode::JMP_IF_FALSE, 0, node.loc);
-
-        this->loopStack.push_back(LoopContext{});
-        this->loopStack.back().continueTarget = loopStart;
-
-        this->emitIterableElement(protocol, seqSlot, idxSlot, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, elemSlot, node.loc);
-
-        if (indexSlot >= 0) {
-            this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-            this->current.get().emit(OpCode::STORE_SLOT, indexSlot, node.loc);
-        }
-
-        int oneIdx = this->addConstant(1);
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-        this->current.get().emit(OpCode::PUSH_INT, oneIdx, node.loc);
-        this->current.get().emit(OpCode::ADD, 0, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, idxSlot, node.loc);
-
-        node.body->accept(*this);
-        this->current.get().emit(OpCode::POP, 0, node.loc);
-
-        size_t jmpBackIdx = this->current.get().emit(OpCode::JMP, 0, node.loc);
-        this->current.get().patchJump(jmpBackIdx, static_cast<int>(loopStart) - static_cast<int>(jmpBackIdx) - 1);
-
-        size_t exitPos = this->current.get().currentIP();
-        this->current.get().patchJump(jmpFalseIdx, static_cast<int>(exitPos) - static_cast<int>(jmpFalseIdx) - 1);
-
-        for (size_t idx : this->loopStack.back().breakJumps)
-            this->current.get().patchJump(idx, static_cast<int>(exitPos) - static_cast<int>(idx) - 1);
-        for (size_t idx : this->loopStack.back().continueJumps)
-            this->current.get().patchJump(idx, static_cast<int>(loopStart) - static_cast<int>(idx) - 1);
-
-        this->loopStack.pop_back();
-
-        int emptyIdx = this->addConstant(std::string(""));
-        this->current.get().emit(OpCode::PUSH_STR, emptyIdx, node.loc);
-    }
-
-    void Compiler::compileForInCounter(ForInNode& node, size_t uid) {
-        auto& range = static_cast<RangeNode&>(*node.iterable);
-
-        const int idxSlot = this->declareSlot("__forin_idx_" + std::to_string(uid));
-        const int endSlot = this->declareSlot("__forin_end_" + std::to_string(uid));
-        const int dirSlot = this->declareSlot("__forin_dir_" + std::to_string(uid));
-        const int elemSlot = this->declareSlot(node.elementVar);
-        const int indexSlot = node.hasIndexVar ? this->declareSlot(node.indexVar) : -1;
-
-        this->compileValue(*range.start, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, idxSlot, node.loc);
-
-        this->compileValue(*range.end, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, endSlot, node.loc);
-
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-        this->current.get().emit(OpCode::LOAD_SLOT, endSlot, node.loc);
-        this->current.get().emit(OpCode::CMP_LE, 0, node.loc);
-
-        size_t jmpDescIdx = this->current.get().emit(OpCode::JMP_IF_FALSE, 0, node.loc);
-
-        int ascIdx = this->addConstant(1);
-        this->current.get().emit(OpCode::PUSH_INT, ascIdx, node.loc);
-
-        size_t jmpDirEndIdx = this->current.get().emit(OpCode::JMP, 0, node.loc);
-
-        int descStart = static_cast<int>(this->current.get().currentIP());
-        this->current.get().patchJump(jmpDescIdx, descStart - static_cast<int>(jmpDescIdx) - 1);
-
-        int descIdx = this->addConstant(-1);
-        this->current.get().emit(OpCode::PUSH_INT, descIdx, node.loc);
-
-        int dirEndPos = static_cast<int>(this->current.get().currentIP());
-        this->current.get().patchJump(jmpDirEndIdx, dirEndPos - static_cast<int>(jmpDirEndIdx) - 1);
-
-        this->current.get().emit(OpCode::STORE_SLOT, dirSlot, node.loc);
-
-        size_t loopStart = this->current.get().currentIP();
-
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-        this->current.get().emit(OpCode::LOAD_SLOT, endSlot, node.loc);
-        this->current.get().emit(OpCode::SUB, 0, node.loc);
-        this->current.get().emit(OpCode::LOAD_SLOT, dirSlot, node.loc);
-        this->current.get().emit(OpCode::MUL, 0, node.loc);
-        int zeroIdx = this->addConstant(0);
-        this->current.get().emit(OpCode::PUSH_INT, zeroIdx, node.loc);
-        this->current.get().emit(OpCode::CMP_LT, 0, node.loc);
-
-        size_t jmpFalseIdx = this->current.get().emit(OpCode::JMP_IF_FALSE, 0, node.loc);
-
-        this->loopStack.push_back(LoopContext{});
-        this->loopStack.back().continueTarget = loopStart;
-
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, elemSlot, node.loc);
-
-        if (indexSlot >= 0) {
-            this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-            this->current.get().emit(OpCode::STORE_SLOT, indexSlot, node.loc);
-        }
-
-        this->current.get().emit(OpCode::LOAD_SLOT, idxSlot, node.loc);
-        this->current.get().emit(OpCode::LOAD_SLOT, dirSlot, node.loc);
-        this->current.get().emit(OpCode::ADD, 0, node.loc);
-        this->current.get().emit(OpCode::STORE_SLOT, idxSlot, node.loc);
-
-        node.body->accept(*this);
-        this->current.get().emit(OpCode::POP, 0, node.loc);
-
-        size_t jmpBackIdx = this->current.get().emit(OpCode::JMP, 0, node.loc);
-        this->current.get().patchJump(jmpBackIdx, static_cast<int>(loopStart) - static_cast<int>(jmpBackIdx) - 1);
-
-        size_t exitPos = this->current.get().currentIP();
-        this->current.get().patchJump(jmpFalseIdx, static_cast<int>(exitPos) - static_cast<int>(jmpFalseIdx) - 1);
-
-        for (size_t idx : this->loopStack.back().breakJumps)
-            this->current.get().patchJump(idx, static_cast<int>(exitPos) - static_cast<int>(idx) - 1);
-        for (size_t idx : this->loopStack.back().continueJumps)
-            this->current.get().patchJump(idx, static_cast<int>(loopStart) - static_cast<int>(idx) - 1);
-
-        this->loopStack.pop_back();
-
-        int emptyIdx = this->addConstant(std::string(""));
-        this->current.get().emit(OpCode::PUSH_STR, emptyIdx, node.loc);
     }
 
     void Compiler::emitArithmeticOp(const std::string& op, const SourceLocation& loc) {
@@ -766,191 +565,6 @@ namespace LOICollection::frontend::ir {
     void Compiler::visit(ClassNode& node) {
         this->registerClassMeta(node);
         this->compileClassBodies(node);
-    }
-
-    void Compiler::registerClassMeta(ClassNode& node) {
-        if (classIndices.contains(node.name))
-            return;
-
-        int baseIdx = -1;
-        if (!node.baseClassName.empty()) {
-            if (this->registeringClasses.contains(node.name)) {
-                this->diagnostics.addError(node.loc,
-                    "Circular inheritance involving class '" + node.name + "'");
-                return;
-            }
-
-            this->registeringClasses.insert(node.name);
-
-            auto baseIt = this->classNodes.find(node.baseClassName);
-            if (baseIt == this->classNodes.end()) {
-                this->diagnostics.addError(node.loc, "Unknown base class: " + node.baseClassName);
-                this->registeringClasses.erase(node.name);
-                return;
-            }
-
-            if (!this->classIndices.contains(node.baseClassName))
-                this->registerClassMeta(baseIt->second.get());
-
-            auto registeredIt = this->classIndices.find(node.baseClassName);
-            if (registeredIt == this->classIndices.end()) {
-                this->diagnostics.addError(node.loc, "Failed to register base class: " + node.baseClassName);
-                this->registeringClasses.erase(node.name);
-                return;
-            }
-
-            baseIdx = registeredIt->second;
-            this->registeringClasses.erase(node.name);
-        }
-
-        int classIdx = static_cast<int>(this->chunk.classes.size());
-        this->classIndices[node.name] = classIdx;
-        this->registeredClasses.insert(classIdx);
-
-        ir::ClassMeta meta;
-        meta.name = node.name;
-        meta.baseClassIndex = baseIdx;
-
-        if (baseIdx >= 0) {
-            const auto& base = this->chunk.classes[baseIdx];
-            meta.fieldNames = base.fieldNames;
-            meta.defaults = base.defaults;
-            meta.hasDefault = base.hasDefault;
-            meta.constructorIndex = base.constructorIndex;
-            meta.methods = base.methods;
-            meta.methodSignatures = base.methodSignatures;
-            meta.staticMethods = base.staticMethods;
-            meta.staticMethodSignatures = base.staticMethodSignatures;
-            meta.staticFieldNames = base.staticFieldNames;
-            meta.staticDefaults = base.staticDefaults;
-            meta.staticHasDefault = base.staticHasDefault;
-            meta.ancestorIndices.push_back(baseIdx);
-            meta.ancestorIndices.insert(meta.ancestorIndices.end(), base.ancestorIndices.begin(), base.ancestorIndices.end());
-        }
-
-        for (const auto& member : node.members) {
-            auto& names = member.isStatic ? meta.staticFieldNames : meta.fieldNames;
-            auto& defaults = member.isStatic ? meta.staticDefaults : meta.defaults;
-            auto& hasDefault = member.isStatic ? meta.staticHasDefault : meta.hasDefault;
-
-            auto fieldIt = std::ranges::find(names, member.name);
-            if (fieldIt == names.end()) {
-                names.push_back(member.name);
-                hasDefault.push_back(member.hasDefault);
-                defaults.emplace_back();
-                fieldIt = std::ranges::find(names, member.name);
-            } else {
-                auto fieldIdx = static_cast<size_t>(std::distance(names.begin(), fieldIt));
-                hasDefault[fieldIdx] = member.hasDefault;
-            }
-
-            auto fieldIdx = static_cast<size_t>(std::distance(names.begin(), fieldIt));
-            if (member.hasDefault) {
-                if (member.defaultExpr) {
-                    auto value = this->constantValue(*member.defaultExpr);
-                    if (value.has_value()) {
-                        defaults[fieldIdx] = *value;
-                        continue;
-                    }
-                }
-
-                this->diagnostics.addError(node.loc,
-                    "Member default value of '" + member.name + "' must be a constant literal");
-                defaults[fieldIdx] = ValueNode::ValueType{};
-            } else {
-                defaults[fieldIdx] = ValueNode::ValueType{};
-            }
-        }
-
-        for (auto & method : node.methods) {
-            int methodIdx = static_cast<int>(methodCount++);
-
-            if (method.isConstructor)
-                meta.constructorIndex = methodIdx;
-            else {
-                std::string signature = this->methodSignature(method);
-                auto& signatures = method.isStatic ? meta.staticMethodSignatures : meta.methodSignatures;
-                auto& methods = method.isStatic ? meta.staticMethods : meta.methods;
-
-                auto sigIt = std::ranges::find(signatures, signature);
-                if (sigIt != signatures.end()) {
-                    auto ordinal = static_cast<size_t>(std::distance(signatures.begin(), sigIt));
-                    methods[ordinal] = methodIdx;
-                } else {
-                    signatures.push_back(signature);
-                    methods.push_back(methodIdx);
-                }
-            }
-        }
-
-        this->classMethodIndices[node.name] = meta.methods;
-        this->classStaticMethodIndices[node.name] = meta.staticMethods;
-        this->chunk.classes.push_back(std::move(meta));
-        this->bodyOrder.emplace_back(std::ref(node));
-    }
-
-    void Compiler::compileClassBodies(ClassNode& node) {
-        auto classIt = this->classIndices.find(node.name);
-        if (classIt == this->classIndices.end() || !this->registeredClasses.contains(classIt->second))
-            return;
-
-        int classIdx = classIt->second;
-
-        for (const auto& method : node.methods) {
-            ir::MethodMeta mm;
-            mm.name = method.name;
-            mm.classIndex = classIdx;
-            mm.argCount = static_cast<int>(method.params.size());
-
-            int bodyIdx = static_cast<int>(this->chunk.methodBodies.size());
-            auto body = std::make_unique<BytecodeChunk>();
-            BytecodeChunk& bodyChunk = *body;
-            this->chunk.methodBodies.push_back(std::move(body));
-
-            std::reference_wrapper<BytecodeChunk> saved = this->current;
-            this->current = std::ref(bodyChunk);
-            auto loops = this->suspendLoops();
-
-            this->pushScope(true, 0);
-            for (const auto& param : method.params)
-                this->declareSlot(param.name);
-
-            if (method.isConstructor && !node.baseClassName.empty() && !method.hasSuperCall) {
-                int ctorIdx = -1;
-                auto baseIt = this->classIndices.find(node.baseClassName);
-                int walkIdx = baseIt == this->classIndices.end() ? -1 : baseIt->second;
-                while (walkIdx >= 0) {
-                    const auto& walkCls = this->chunk.classes[walkIdx];
-                    if (walkCls.constructorIndex != -1) {
-                        ctorIdx = walkCls.constructorIndex;
-                        break;
-                    }
-                    walkIdx = walkCls.baseClassIndex;
-                }
-
-                int argCount = ctorIdx >= 0 ? this->chunk.methods[ctorIdx].argCount : 0;
-                int superIdx = this->addSuperCall(ctorIdx, argCount);
-                this->current.get().emit(OpCode::LOAD_THIS, 0, node.loc);
-                this->current.get().emit(OpCode::CALL_SUPER_CTOR, superIdx, node.loc);
-                this->current.get().emit(OpCode::POP, 0, node.loc);
-            }
-
-            if (method.body)
-                method.body->accept(*this);
-
-            this->current.get().emit(OpCode::POP);
-
-            int emptyIdx = this->addConstant(std::string(""));
-            this->current.get().emit(OpCode::PUSH_STR, emptyIdx);
-            this->current.get().emit(OpCode::RETURN);
-
-            this->current = saved;
-            bodyChunk.slotCount = this->closeScope();
-
-            mm.bodyIndex = bodyIdx;
-
-            this->chunk.methods.push_back(std::move(mm));
-        }
     }
 
     void Compiler::visit(ReturnNode& node) {
@@ -1559,18 +1173,6 @@ namespace LOICollection::frontend::ir {
         };
     }
 
-    std::string Compiler::methodSignature(const MethodDecl& method) const {
-        std::string signature = method.name + "(";
-        for (size_t i = 0; i < method.paramTypes.size(); ++i) {
-            if (i != 0)
-                signature += ",";
-
-            signature += typeInfoToString(method.paramTypes[i]);
-        }
-        signature += ")";
-        return signature;
-    }
-
     void Compiler::compileValue(ExprNode& node, const SourceLocation& loc) {
         node.accept(*this);
 
@@ -1600,4 +1202,5 @@ namespace LOICollection::frontend::ir {
 
         return std::nullopt;
     }
+
 }
